requirements.txt
flask
ollama
pypdf2
torch
transformers
ultralytics
pillow
pytesseract